api connected to the ui

+++ app/api/posts/[id]/route.ts
@@ -1,5 +1,7 @@
 import prisma from "@/lib/prismadb";
+import {getServerSession} from "next-auth";
 import {NextResponse} from "next/server";
+import {authOptions} from "../../auth/[...nextauth]/route";
 
 export const GET = async (req: Request, {params}: any) => {
   try {
@@ -12,6 +14,12 @@ export const GET = async (req: Request, {params}: any) => {
 };
 
 export const PUT = async (req: Request, {params}: any) => {
+  const session = await getServerSession(authOptions);
+
+  if (!session) {
+    return NextResponse.json({error: "Not Authenticated"}, {status: 401});
+  }
+
   const {title, content, links, selectedCategory, imageUrl, publicId} =
     await req.json();
   const id = params.id;
@@ -34,6 +42,11 @@ export const PUT = async (req: Request, {params}: any) => {
 };
 
 export const DELETE = async (req: Request, {params}: any) => {
+  const session = await getServerSession(authOptions);
+
+  if (!session) {
+    return NextResponse.json({error: "Not Authenticated"}, {status: 401});
+  }
   const id = params.id;
   try {
     const post = await prisma.post.delete({where: {id}});
